Split concerns: queries, mutations, tests
